Add files via upload
Due to crater and boulder distribution on the surface of the moon, altitude and vertical velocity information often is difficult to obtain. Satellite radar altimeter has a potential to monitor height and velocity variation. Here, it is shown that the radar altimeter can successfully track the altitude and vertical velocity from the height 9.9 km to 2 m. For the landing of Chandrayaan-2 on the Lunar surface, the data obtained by radar altimeter is very crucial. This paper describes the altimeter operation principle, H/w data acquisition, ADC analysis, altitude and Vertical Velocity results.