Add files via upload
Festivals are an integral part of Indian culture. The traditions associated with the celebrations of these festivals reflect people’s deep rooted faith in deities. At the same time people are ignorant of the adversities they cause to the environment. For instance, the festival of Ganesh Chaturthi, Durga Ashthmi are celebrated in different parts of India. On the last day of the festival the idols are taken out in a colorful and musical procession and traditionally immersed in water bodies. Since immersion is a community activity the place chosen is the one easily accessible like a pond, stream, river or sea. Most of these idols are made of plaster of paris and harmful chemicals. The Plaster of Paris does not easily dissolve in water. This results in the reduction of oxygen level in water. Apart, chemical paints used for decorating idols are hazard to marine life and people. These chemicals contain heavy metals and are bio-accumulative. These harmful substances get dissolved in water bodies in due course of time posing heavy threat to human life by negatively affecting the food chain. This paper focusses on the alternatives to plaster of paris and chemicals. It also reflects on the problems of the artisans, the government schemes and the need for adopting these alternatives so as to take a positive step towards sustainable environment.